+++ code/MPI/read_coo_matrix.h
@@ -7,7 +7,7 @@ typedef struct
 {
 	int* row_idx;
 	int* col_idx;
-	double* val;
+	float* val;
 	int nnz;
 }coo_mat;
 
@@ -36,7 +36,7 @@ coo_mat* read_file(char* input_filename)
 
 	mat_read->row_idx = (int*)malloc(nnz*sizeof(int));
 	mat_read->col_idx = (int*)malloc(nnz*sizeof(int));
-	mat_read->val     = (double*)malloc(nnz*sizeof(double));
+	mat_read->val     = (float*)malloc(nnz*sizeof(float));
 	mat_read->nnz     = nnz;
 
 	r8st_data_read (input_filename, row, col, nnz, mat_read->row_idx, mat_read->col_idx, mat_read->val);
